requirements.txt
absl-py==2.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==24.2.0
audioread==3.0.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4 @ file:///home/conda/feedstock_root/build_artifacts/beautifulsoup4_1649463573192/work
black==23.3.0
bleach==6.0.0
blinker==1.8.2
blis==0.7.9
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1666764778428/work
catalogue==2.0.8
cattrs==23.2.3
certifi==2024.7.4
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1666184005339/work
cfgv==3.4.0
chardet==3.0.4
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
click==8.1.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm==0.1.2
conda==22.11.1
conda-content-trust @ file:///home/conda/feedstock_root/build_artifacts/conda-content-trust_1621370699668/work
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1669907009957/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1669733752472/work
confection==0.0.4
contourpy==1.0.7
coremltools==7.2
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1669593291283/work
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.7
datasets==2.11.0
debugpy==1.6.6
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
deprecation==2.0.7
dill==0.3.6
distlib==0.3.7
distro==1.9.0
et-xmlfile==1.0.1
exceptiongroup==1.1.3
executing==1.2.0
fastai==2.7.12
fastbook==0.0.29
fastcore==1.5.29
fastdownload==0.0.7
fastjsonschema==2.16.2
fastprogress==1.0.3
filelock==3.12.2
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.4.0
fuzzywuzzy==0.18.0
gast==0.6.0
google-pasta==0.2.0
googletrans==4.0.0rc1
gpt-engineer==0.0.8
graphviz==0.20.1
greenlet @ file:///Users/runner/miniforge3/conda-bld/greenlet_1667836816933/work
grpcio==1.66.0
h11==0.9.0
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
hstspreload==2024.5.1
html5lib @ file:///home/conda/feedstock_root/build_artifacts/html5lib_1592930327044/work
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.12.0
hyperframe==5.2.0
identify==2.5.26
idna==2.10
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==7.7.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.5.0
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.4
kaggle==1.6.17
keras==3.5.0
kiwisolver==1.4.4
langchain==0.0.267
langcodes==3.3.0
langsmith==0.0.24
lazy_loader==0.2
Levenshtein==0.21.1
libclang==18.1.1
librosa==0.10.0.post2
llvmlite==0.39.1
lxml @ file:///Users/runner/miniforge3/conda-bld/lxml_1671013714685/work
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.20.1
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
ml-dtypes==0.4.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mypy==1.3.0
mypy-extensions==1.0.0
namex==0.0.8
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
nltk==3.8.1
nodeenv==1.8.0
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numexpr==2.8.5
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1668919183251/work
openai==0.28.0
openapi-schema-pydantic==1.2.4
openpyxl==3.0.10
opt-einsum==3.3.0
optree==0.12.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1670530880680/work
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.10.0
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1667232663820/work
pooch==1.6.0
pre-commit==3.3.3
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==24.7.0
pyarrow==11.0.0
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1666836597919/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydantic==1.10.7
pydub==0.25.1
Pygments==2.14.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1665350324128/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-json-logger==2.0.4
python-Levenshtein==0.21.1
python-slugify==8.0.4
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
rapidfuzz==3.1.1
regex==2023.3.23
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1661872987712/work
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.1
ruamel-yaml-conda @ file:///Users/runner/miniforge3/conda-bld/ruamel_yaml_1666819777457/work
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1666827345171/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1670412837369/work
rudder-sdk-python==2.0.2
ruff==0.0.272
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.1
Send2Trash==1.8.0
sentencepiece==0.1.98
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==6.3.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve @ file:///home/conda/feedstock_root/build_artifacts/soupsieve_1658207591808/work
soxr==0.3.5
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy @ file:///Users/runner/miniforge3/conda-bld/sqlalchemy_1670718343328/work
srsly==2.4.6
stack-data==0.6.2
sympy==1.13.2
tables @ file:///Users/runner/miniforge3/conda-bld/pytables_1666850067546/work
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
thinc==8.1.9
threadpoolctl==3.1.0
tiktoken==0.4.0
timm==0.6.12
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1657485559105/work
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1662214488106/work
traitlets==5.9.0
transformers==4.28.1
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.9.0
uri-template==1.2.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1669259737463/work
virtualenv==20.24.3
wasabi==1.1.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==3.0.4
widgetsnbextension==3.6.4
wrapt==1.16.0
xlrd @ file:///home/conda/feedstock_root/build_artifacts/xlrd_1610224409810/work
xxhash==3.2.0
yarl==1.9.2
zipp==3.13.0
zstandard==0.19.0
